break out ordered-map-model.ts from metamodel.ts
