+++ lib/js/metamodel/metamodel.ts
@@ -17,7 +17,6 @@ import {
     keyConstraintError,
     immutableWriteError,
     isDraftKeyError,
-    _serializeContainer,
 } from './base-model.ts';
 
 // These are the exports from ./base-model.ts that are used beyond this
@@ -49,10 +48,10 @@ export {
 import {
     _NOTDEF,
     iterMap,
-    sort_alpha,
     objectEntriesAreEqual,
     collectDependencies,
     unwrapPotentialWriteProxy,
+    PATH_SEPARATOR,
 } from './util.ts';
 
 export { objectEntriesAreEqual, collectDependencies };
@@ -105,7 +104,6 @@ export {
 import {
     IS_WRAPPER_TYPE,
     _LOCAL_PROXIES,
-    _OLD_TO_NEW_SLOT,
     _HAS_DRAFT_FOR_PROXY,
     _HAS_DRAFT_FOR_OLD_STATE_KEY,
     _GET_DRAFT_FOR_PROXY,
@@ -146,860 +144,8 @@ export {
 import { _AbstractListModel } from './list-model.ts';
 export { _AbstractListModel };
 
-
-
-/**
- * The only use of this function so far is to demo custom ordering in
- * _AbstractOrderedMapModel. It's not included as a fixed option,
- * like e.g. KEYS_ALPHA, as it doesn't really have an actual use-case IMO.
- *
- * setup = {
- *      ordering: _AbstractOrderedMapModel.ORDER.CUSTOM
- *    , customOrderingFn: toShuffle
- * }
- * const MyModelMap = _AbstractOrderedMapModel.createClass('MyModelMap', MyModel, setup);
- */
-export function toShuffle(from) {
-    const array = Array.from(from);
-    let currentIndex = array.length;
-    // While there remain elements to shuffle...
-    while (currentIndex != 0) {
-        // Pick a remaining element...
-        const randomIndex = Math.floor(Math.random() * currentIndex);
-        currentIndex--;
-        // And swap it with the current element.
-        [array[currentIndex], array[randomIndex]] = [
-            array[randomIndex],
-            array[currentIndex],
-        ];
-    }
-    return array;
-}
-
-// combines the order and inserting logic of the _AbstractListModel
-// with the uniqueness by the keys of the _AbstractMapModel
-// FIXME it should be possible to have a way to validate keys
-//       that is attached to the concrete class via createClass.
-const MAP_ORDER = Object.freeze({
-    KEYS_ALPHA: Symbol("ORDER_KEYS_ALPHA"),
-    KEYS_ALPHA_REVERSE: Symbol("ORDER_KEYS_ALPHA_REVERSE"),
-    CUSTOM: Symbol("ORDER_CUSTOM"),
-    CUSTOM_REVERSE: Symbol("ORDER_CUSTOM_REVERSE"),
-});
-
-export class _AbstractOrderedMapModel extends _BaseContainerModel {
-    static Model: _BaseModel;
-
-    static get dependencies() {
-        return this.Model.dependencies;
-    }
-
-    static ORDER = MAP_ORDER;
-
-    static [MAP_ORDER.KEYS_ALPHA](values) {
-        const entries = new Map(),
-            keys = [];
-        for (const [key, data] of values) {
-            entries.set(key, [key, data]);
-            keys.push(key);
-        }
-        return keys.sort(sort_alpha).map((key) => entries.get(key));
-    }
-
-    static [MAP_ORDER.KEYS_ALPHA_REVERSE](values) {
-        return this[this.ORDER.KEYS_ALPHA](values).reverse();
-    }
-
-    static [MAP_ORDER.CUSTOM](values) {
-        return this._customOrderingFn(values);
-    }
-
-    static [MAP_ORDER.CUSTOM_REVERSE](values) {
-        return this._customOrderingFn(values).reverse();
-    }
-
-    static createClass(className: string, Model: _BaseModel, setup = {}) {
-        // jshint unused: vars
-        setup = {
-            ordering: _NOTDEF,
-            customOrderingFn: _NOTDEF,
-            validateKeyFn: _NOTDEF,
-            ...setup,
-        };
-        if (setup.ordering !== _NOTDEF) {
-            const availableOrderingSymbols = new Set(Object.values(this.ORDER));
-            if (!availableOrderingSymbols.has(setup.ordering))
-                throw new Error(
-                    `KEY ERROR setup.ordering unknown "${setup.ordering.toString()}" ` +
-                        `allowed values are ${this.name}.ORDER.(${Object.keys(this.ORDER).join("|")}).`,
-                );
-            if (
-                setup.ordering === this.ORDER.CUSTOM ||
-                setup.ordering === this.ORDER.CUSTOM_REVERSE
-            ) {
-                if (setup.customOrderingFn === _NOTDEF)
-                    throw new Error(
-                        `VALUE ERROR setup.ordering is "${setup.ordering.toString()}" but setup.customOrderingFn is not specified`,
-                    );
-                if (typeof setup.customOrderingFn !== "function")
-                    throw new Error(
-                        `VALUE ERROR setup.customOrderingFn must be a function but is "${typeof setup.customOrderingFn}".`,
-                    );
-                // NOTE: Not checking here if that function behaves correctly!
-            }
-        }
-        if (setup.validateKeyFn !== _NOTDEF) {
-            if (typeof setup.validateKeyFn !== "function")
-                throw new Error(
-                    `VALUE ERROR setup.validateKeyFn must be a function but is "${typeof setup.validateKeyFn}".`,
-                );
-        }
-
-        // this way name will naturally become class.name.
-        const result = {
-            // jshint ignore: start
-            [className]: class extends this {
-                static Model: _BaseModel = Model;
-                static ORDERING =
-                    setup.ordering === _NOTDEF ? null : setup.ordering;
-                static _customOrderingFn =
-                    setup.customOrderingFn === _NOTDEF
-                        ? function (/*values*/) {
-                              throw new Error(
-                                  `NOT IMPLEMENTED _customOrderingFn is not defined in ${this.name}`,
-                              );
-                          }
-                        : setup.customOrderingFn;
-                static validateKeyFn =
-                    setup.validateKeyFn === _NOTDEF
-                        ? null
-                        : setup.validateKeyFn;
-            },
-            // jshint ignore: end
-        };
-        // Can't override class.Model anymore, would be possible w/o the freeze.
-        // Maybe only fix "Model" property?
-        // Object.freeze(result[className]);
-        return result[className];
-    }
-
-    constructor(
-        oldState = null,
-        dependencies = null,
-        serializedValue = null,
-        serializeOptions = SERIALIZE_OPTIONS,
-    ) {
-        if (oldState === null && dependencies === null)
-            throw new Error(
-                `TYPE ERROR either oldState or dependencies are required.`,
-            );
-        if (oldState !== null && dependencies !== null)
-            // The problem is that metamorphose may return OLD_STATE but
-            // as a constructor that is invoked with the `new` keyword,
-            // a new object must be returned.
-            throw new Error(
-                `TYPE ERROR can't constuct with both oldState and dependencies`,
-            );
-
-        if (oldState && oldState.isDraft)
-            throw new Error(
-                `LIFECYCLE ERROR ` +
-                    `oldState ${oldState} is draft but must be immutable.`,
-            );
-        super(oldState);
-
-        // Start with an empty this._value for quick not-changed comparison.
-        Object.defineProperty(this, "_value", {
-            value: new Array(
-                this[OLD_STATE] !== null ? this[OLD_STATE].length : 0,
-            ),
-            writable: false, // can't replace the array itself
-            configurable: true,
-        });
-        Object.defineProperty(this, "_keys", {
-            value: new FreezableMap(),
-            writable: false, // can't replace the FreezableMap itself
-            configurable: true,
-        });
-        Object.defineProperty(this, "dependencies", {
-            get: () => {
-                if (this[OLD_STATE] === null)
-                    throw new Error("Primal State has no dependencies yet!");
-                // In draft-mode, this[OLD_STATE] has the dependencies.
-                return this[OLD_STATE].dependencies;
-            },
-            configurable: true,
-        });
-        // Keep track of proxies and OLD_STATE original indexes in a
-        // shadow of this._value that is kept in sync with value!
-        // Entries may get replaced by set or moved/removed by splice.
-
-        this[_OLD_TO_NEW_SLOT] =
-            this[OLD_STATE] !== null
-                ? [...this[OLD_STATE]].map(([key /*value*/], index) => [
-                      index,
-                      key,
-                      null /*proxy*/,
-                  ])
-                : [];
-        this._updateKeys();
-
-        // Create an immutable primal state if OLD_STATE is null:
-        if (dependencies !== null) {
-            if (serializedValue !== null) {
-                // I don't want to have this as an direct argument
-                // of metamorphose so far, as this way it's made sure
-                // that only the Primal State can load the serialized
-                // state. If it's not a primary state, change is handled
-                // differently.
-                this[_PRIMARY_SERIALIZED_VALUE] = [
-                    serializedValue,
-                    serializeOptions,
-                ];
-            }
-            // Must return a new object (when called with `new`).
-            // only works when there was no OLD_STATE
-            if (dependencies !== _DEFERRED_DEPENDENCIES)
-                return this.metamorphose(dependencies);
-        }
-    }
-
-    *#_metamorphoseGen(dependencies = {}) {
-        //CAUTION: `this` is the object not the class.
-        //
-        // All the following runs, to change deep down a single axis location value.
-        //        It is very important to make this as lightweight as possible.
-        //        At the same time, it is important that the change bubbles
-        //        through the whole structure.
-
-        // Allow case without or with incomplete dependencies argument,
-        // will re-use this[OLD_STATE].dependencies.
-        // Fails in the if dependencies are missing.
-        const dependenciesData = collectDependencies(
-            this.constructor.dependencies,
-            dependencies,
-            this[OLD_STATE]?.dependencies,
-        );
-
-        Object.defineProperty(this, "dependencies", {
-            value: Object.freeze(dependenciesData),
-            writable: true,
-            configurable: true,
-        });
-
-        if (this[_PRIMARY_SERIALIZED_VALUE]) {
-            const [serializedValues, serializeOptions] =
-                    this[_PRIMARY_SERIALIZED_VALUE],
-                childItems = [];
-            for (const [key, serializedValue] of serializedValues) {
-                const childItem =
-                    yield* this.constructor.Model.createPrimalStateGen(
-                        this.dependencies,
-                        serializedValue,
-                        serializeOptions,
-                    );
-                childItems.push([key, childItem]);
-            }
-            this.push(...childItems);
-        }
-        // Don't keep this
-        delete this[_PRIMARY_SERIALIZED_VALUE];
-
-        const dependenciesAreEqual =
-            this[OLD_STATE] !== null &&
-            objectEntriesAreEqual(
-                this[OLD_STATE].dependencies,
-                this.dependencies,
-            );
-
-        // shortcut
-        if (
-            dependenciesAreEqual &&
-            this.size === this[OLD_STATE].size &&
-            // is only empty slots i.e. no changes
-            Object.values(this._value).length === 0
-        )
-            return this[OLD_STATE];
-
-        for (const index of this._value.keys()) {
-            const kvItem = Object.hasOwn(this._value, index)
-                ? this._value[index]
-                : this[OLD_STATE].value[this[_OLD_TO_NEW_SLOT][index][0]];
-            const [key, item] = kvItem || [];
-
-            if (!(item instanceof this.constructor.Model))
-                throw new Error(
-                    `TYPE ERROR ${this.constructor.name} ` +
-                        `expects ${this.constructor.Model.name} ` +
-                        `wrong type at ${key} in ${index} ("${item}" typeof ${typeof item}).`,
-                );
-            const immutable = item.isDraft
-                ? yield* item.metamorphoseGen(this.dependencies)
-                : // Not sure if we should check with objectEntriesAreEqual
-                  // or just let entry check itself if it has to move forward.
-                  !objectEntriesAreEqual(this.dependencies, item.dependencies)
-                  ? yield* item.getDraft().metamorphoseGen(this.dependencies)
-                  : item;
-            this._value[index] = Object.freeze([key, immutable]);
-        }
-
-        // NOTE: this only ensures that the immutable map is ordered
-        // the mutable map can be in undefined order.
-        if (this.constructor.ORDERING !== null) {
-            // this._keys will get updated in #_lockAndFreeze via _updateKeys();
-            const newlyOrderedEntries = this.constructor[
-                this.constructor.ORDERING
-            ](this._value);
-            this._value.splice(0, Infinity);
-            for (const entry of newlyOrderedEntries) this._value.push(entry);
-        }
-
-        // last stop to detect a no-change
-        if (
-            this[OLD_STATE] !== null &&
-            dependenciesAreEqual &&
-            this.size === this[OLD_STATE].size &&
-            this._value.every((entry, index) => {
-                const [key, value] = entry,
-                    [oldKey, oldValue] = this[OLD_STATE].value[index];
-                return key === oldKey && value === oldValue;
-            })
-        )
-            return this[OLD_STATE];
-        return this;
-    }
-
-    #_lockAndFreeze() {
-        Object.defineProperty(this, "dependencies", {
-            value: Object.freeze(this.dependencies),
-            writable: true,
-            configurable: true,
-        });
-        delete this[OLD_STATE];
-        Object.defineProperty(this, "_value", {
-            value: Object.freeze(this._value),
-            writable: false,
-            configurable: false,
-        });
-        this._updateKeys();
-        Object.defineProperty(this, "_keys", {
-            value: Object.freeze(this._keys),
-            writable: false,
-            configurable: false,
-        });
-
-        Object.defineProperty(this, _IS_DRAFT_MARKER, {
-            value: false,
-            configurable: false,
-        });
-        delete this[_OLD_TO_NEW_SLOT];
-        Object.freeze(this);
-    }
-
-    #_metamorphoseCleanUp() {
-        delete this.dependencies;
-    }
-
-    *metamorphoseGen(dependencies = {}) {
-        if (!this.isDraft)
-            throw new Error(
-                `LIFECYCLE ERROR ${this} must be in draft mode to metamorphose.`,
-            );
-        let result;
-        try {
-            result = yield* this.#_metamorphoseGen(dependencies);
-        } finally {
-            if (result === this) {
-                // This metamorphosed into a new state!
-                this.#_lockAndFreeze();
-            } else {
-                // on error or if(result === this[OLD_STATE]) {
-                // reset metamorphose residues so that this draft could
-                // be metamorphosed again (it happens, see the commit).
-                this.#_metamorphoseCleanUp();
-            }
-        }
-        return result;
-    }
-
-    _updateKeys() {
-        this._keys.clear();
-        for (const index of this._value.keys()) {
-            const key = Object.hasOwn(this._value, index)
-                ? this._value[index][0]
-                : this[_OLD_TO_NEW_SLOT][index][1];
-            this._keys.set(key, index);
-        }
-    }
-
-    get value() {
-        if (this.isDraft)
-            throw new Error(
-                `LIFECYCLE ERROR ${this} will only return this.value when immutable/not a draft..`,
-            );
-        return this._value;
-    }
-
-    *[Symbol.iterator]() {
-        for (const key of this.ownKeys()) yield [key, this.get(key)];
-    }
-
-    *indexedEntries() {
-        for (const [key, value] of this) {
-            const [index /* error message*/] = this.keyToIndex(key);
-            yield [index, [key, value]];
-        }
-    }
-
-    get length() {
-        return this._value.length;
-    }
-
-    get size() {
-        return this._value.length;
-    }
-
-    hasOwn(key) {
-        return this._keys.has(key);
-    }
-    /**
-     * hasDraftFor and getDraftFor are likely only a required interfaces
-     * for _BaseContainerModel.
-     */
-    [_HAS_DRAFT_FOR_PROXY](proxy) {
-        if (!this.isDraft) return false;
-
-        const index = this[_OLD_TO_NEW_SLOT].findIndex(
-            ([, , ownProxy]) => ownProxy === proxy,
-        );
-        if (index === -1)
-            // proxy is disconnected
-            return false;
-
-        // May not be a draft at this moment!
-        // In that case it may also not yet be in this._value.
-        const [, /*key*/ item] = (Object.hasOwn(this._value, index) &&
-            this._value[index]) || [null, null];
-        if (!item || !item.isDraft) return false;
-
-        // Item is a draft created here for proxy. We know because
-        // the proxy was used to find the index.
-        return true;
-    }
-    [_HAS_DRAFT_FOR_OLD_STATE_KEY](oldKey) {
-        if (!this.isDraft) return false;
-        const [oldIndex, message] = this[OLD_STATE].keyToIndex(oldKey);
-        if (oldIndex === null)
-            // was not in OLD_STATE
-            throw new Error(message);
-
-        const index = this[_OLD_TO_NEW_SLOT].findIndex(
-            ([, ownOldKey]) => ownOldKey === oldKey,
-        );
-        if (index === -1)
-            // proxy is disconnected
-            return false;
-
-        // May not be a draft at this moment!
-        // In that case it may also not yet be in this._value.
-        const [, /*key*/ item] = (Object.hasOwn(this._value, index) &&
-            this._value[index]) || [null, null];
-        if (!item || !item.isDraft) return false;
-
-        // Item is a draft created here for key. We know because
-        // the key was used to find the index.
-        return true;
-    }
-
-    [_GET_DRAFT_FOR_OLD_STATE_KEY](oldKey) {
-        // key must be in this[OLD_STATE]!
-        // draft will be for this[OLD_STATE].get(key).getDraft()
-        if (!this.isDraft)
-            // required: so this can be turned into a draft on demand
-            throw immutableWriteError(
-                new Error(
-                    `IMMUTABLE WRITE ATTEMPT ` +
-                        `${this}[_GET_DRAFT_FOR_OLD_STATE_KEY](${oldKey}) is immutable, not a draft.`,
-                ),
-            );
-
-        const [oldIndex, message] = this[OLD_STATE].keyToIndex(oldKey);
-        if (oldIndex === null)
-            // was not in OLD_STATE
-            throw new Error(message);
-
-        const index = this[_OLD_TO_NEW_SLOT].findIndex(
-            ([, ownOldKey]) => ownOldKey === oldKey,
-        );
-        if (index === -1)
-            // The item associated with oldIndex is no longer part of this
-            // object, the proxy is disconnected.
-            return false;
-
-        let kvItem = Object.hasOwn(this._value, index) && this._value[index];
-        if (!kvItem) {
-            const item = this[OLD_STATE].get(oldKey);
-            kvItem = [oldKey, item];
-        }
-        const [key, item] = kvItem;
-        if (item.isDraft)
-            // We already created the connection between
-            // index and oldIndex, we found index via oldKey,
-            // item belongs to oldIndex.
-            return item;
-        const draft = item.getDraft();
-        this._value[index] = Object.freeze([key, draft]);
-        return draft;
-    }
-
-    /**
-     * Raises KeyError if key is not available.
-     * Raises ImmutableWriteError if this is not a draft.
-     * Returns False, if draft is not natively created by this function
-     * i.e. set from the outside.
-     * Returns a draft for key otherwise.
-     * this is likely only for _BaseContainerModel
-     */
-    [_GET_DRAFT_FOR_PROXY](proxy) {
-        if (!this.isDraft)
-            throw immutableWriteError(
-                new Error(
-                    `IMMUTABLE WRITE ATTEMPT ` +
-                        `${this} is immutable, not a draft in [_GET_DRAFT_FOR_PROXY].`,
-                ),
-            );
-
-        const index = this[_OLD_TO_NEW_SLOT].findIndex(
-            ([, , ownProxy]) => ownProxy === proxy,
-        );
-        if (index === -1)
-            // proxy is disconnected
-            return false;
-
-        let kvItem = Object.hasOwn(this._value, index) && this._value[index];
-        if (!kvItem) {
-            const [, /*oldIndex*/ key] = this[_OLD_TO_NEW_SLOT][index],
-                // assert key is there, otherwise this will raise a Key Error
-                // also, if oldIndex got removed from _OLD_TO_NEW_SLOT there
-                // must be an item in this._value or the proxy is disconnected.
-                item = this[OLD_STATE].get(key);
-            kvItem = [key, item];
-        }
-        const [key, item] = kvItem;
-        if (item.isDraft)
-            // since we found it via proxy, item belongs to it.
-            // assert this._value[index] === item
-            return item;
-        const draft = item.getDraft();
-        this._value[index] = Object.freeze([key, draft]);
-        return draft;
-    }
-
-    getDraftFor(key, defaultReturn = _NOTDEF) {
-        const proxyOrDraft = this.get(key, defaultReturn);
-        if (_PotentialWriteProxy.isProxy(proxyOrDraft))
-            return this[_GET_DRAFT_FOR_PROXY](proxyOrDraft);
-        return proxyOrDraft;
-    }
-
-    get(key, defaultReturn = _NOTDEF) {
-        const [index, message] = this.keyToIndex(key);
-        if (index === null) {
-            if (defaultReturn !== _NOTDEF) return defaultReturn;
-            throw new Error(message);
-        }
-
-        if (!this.isDraft) return this._value[index][1];
-
-        // Can be a draft or immutable e.g. via set(index, element)
-        let item = Object.hasOwn(this._value, index) && this._value[index][1];
-        if (!item) {
-            // If there's no item in value[index] yet, oldIndex will exist.
-            // FIXME: I guess I could rather just use: this[OLD_STATE].get(key)
-            //        instead of taking this discourse. Of course:
-            //              assert oldKey === key
-            //              assert oldKey === this[OLD_STATE].keyToIndex(key)[0]
-            //        In that case this[_OLD_TO_NEW_SLOT] could be simplified
-            //        as there would be no need to carry oldIndex around!
-            const [oldIndex /*oldKey*/, , proxy] =
-                this[_OLD_TO_NEW_SLOT][index];
-            if (proxy)
-                // Important, otherwise we could create the proxy multiple
-                // times and override the older versions.
-                return proxy;
-            // KeyError if the assumption is wrong, this would require
-            // fixing in here!
-            // Is always immutable.
-            item = this[OLD_STATE].getIndex(oldIndex);
-        }
-
-        // The function understands if item is already a draft
-        // and does not proxify item in that case.
-        const proxyOrDraft = _PotentialWriteProxy.create(this, item);
-        if (_PotentialWriteProxy.isProxy(proxyOrDraft))
-            // it's a proxy
-            this[_OLD_TO_NEW_SLOT][index][2] = proxyOrDraft;
-        // else: It is a draft already and the draft is at this._value[index];
-        return proxyOrDraft;
-    }
-
-    keyToIndex(key) {
-        if (!this._keys.has(key))
-            return [null, `KEY ERROR "${key}" not found.`];
-        return [this._keys.get(key), null];
-    }
-
-    // This method can be handy for the arraySplice method.
-    indexOfKey(key) {
-        return this._keys.has(key) ? this._keys.get(key) : -1;
-    }
-
-    indexToKey(searchIndex) {
-        let index = parseInt(searchIndex, 10);
-        if (isNaN(index))
-            return [null, `KEY ERROR can't parse "${searchIndex}" as integer.`];
-        if (index < 0)
-            // like Array.prototype.at
-            index = index + this._value.length;
-        if (index < 0 || index >= this._value.length)
-            return [
-                null,
-                `KEY ERROR NOT FOUND index "${searchIndex}" is not an index (= ${index})` +
-                    ` (index > 0 && index < ${this._value.length}.`,
-            ];
-
-        const key = Object.hasOwn(this._value, index)
-            ? this._value[index][0]
-            : this[_OLD_TO_NEW_SLOT][index][1];
-        return [key, null];
-    }
-
-    keyOfIndex(index, defaultReturn = _NOTDEF) {
-        const [key, message] = this.indexToKey(index);
-        if (key === null) {
-            if (defaultReturn !== _NOTDEF) return defaultReturn;
-            throw new Error(message);
-        }
-        return key;
-    }
-
-    getIndex(index, defaultReturn = _NOTDEF) {
-        const key = this.keyOfIndex(index, defaultReturn);
-        // via get as the single point of reading
-        // also important so far to have a single point for the proxy
-        // _GET_DRAFT_FOR_OLD_STATE_KEY mechanics, i.e. the result of `get`
-        // will be wrapped into the proxy using the `key` argument.
-        return this.get(key, defaultReturn);
-    }
-
-    ownKeys() {
-        return this._keys.keys();
-    }
-
-    indexOf(item, fromIndex) {
-        // If fromIndex >= array.length, the array is not searched and -1 is returned.
-        if (fromIndex >= this._value.length) return -1;
-
-        // Negative index counts back from the end of the array —
-        // if fromIndex < 0, fromIndex + array.length is used.
-        // Note, the array is still searched from front to back in this case.
-        if (fromIndex < 0) fromIndex = fromIndex + this._value.length;
-
-        // If fromIndex < -array.length or fromIndex is omitted, 0
-        // is used, causing the entire array to be searched.
-        if (fromIndex === undefined || fromIndex < 0) fromIndex = 0;
-
-        const searchArray =
-            fromIndex === 0 ? this._value : this._value.slice(fromIndex);
-        let result = searchArray.findIndex(([, myItem]) => myItem === item);
-        if (result !== -1 && fromIndex) result = result + fromIndex;
-        return result;
-    }
-
-    static validateKey(key) {
-        // FIXME: In theory a lot more types of keys should be allowed
-        // but there's yet no such use case and other types wouldn't
-        // survive serialization
-        const keyType = typeof key,
-            allowedTypes = new Set([typeof "", typeof 0, typeof true]);
-        // Technically true, false, null can be acceptable!
-        // e.g. all of the JSON/JS basic types that serialize
-        // in a unique way. NOTE: null is also a result of
-        // JSON.stringify(NaN) (Infinity) but we only allow null
-        if (
-            (allowedTypes.has(keyType) || key === null) &&
-            (keyType !== "number" || isFinite(key))
-        ) {
-            // FIXME: processing of these keys should be aware that they
-            // are path-parts as a whole and maybe escape them appropriately.
-            // however, this is not implemented yet and "a/key/containing/slashes"
-            // causes touble: "trouble/key" as a key in "/activeState/typeSpec/children/0/stylePatchesSource" triggers:
-            //      Error: KEY ERROR "trouble" not found. (path: /activeState/typeSpec/children/0/stylePatchesSource/trouble/key);
-            //  via StateComparison.getChangedMap
-            if (key.indexOf(Path.SEPARATOR) != -1)
-                return [
-                    false,
-                    `key can't contain the path separator "${Path.SEPARATOR}" ` +
-                        `as path parts are not reliably escaped so far. Key is "${key}". (FIXME)`,
-                ];
-            if (this.validateKeyFn !== null) return this.validateKeyFn(key);
-            return [true, null];
-        }
-        return [
-            false,
-            `the type of a key must be ` +
-                `string, a finite number, boolean or null but key "${key}" is ${keyType}`,
-        ];
-    }
-
-    _validateKey(key) {
-        return this.constructor.validateKey(key);
-    }
-    /**
-     * As a one stop solution, this cleans this._value and rebuilds
-     * all of this._keys. Duplicate keys will be removed.
-     * As entries can override existing keys but also existing keys
-     * can override entries that are inserted before the existing entries
-     * with the same keys, this can be a bit complex.
-     * In general, this will keep the keys that end up later in the
-     * array after insertion and remove the others.
-     * Similar like Object.fromEntries([['a',1], ['a', 2], ['a', 3]])
-     * will create: {'a': 3}.
-     */
-    arraySplice(index, deleteCount, ...entries) {
-        if (!this.isDraft)
-            // FIXME: for the potential write proxy, it becomes very
-            // interesting trying to write many entries.
-            // Also interesting for that when trying to write no entries and just removing stuff.
-            throw immutableWriteError(
-                new Error(
-                    `NOT DRAFT ERROR: ${this} can't call arraySplice when not in draft phase.`,
-                ),
-            );
-
-        const _entries = entries.map((kv, i) => {
-                // Also creates a defensive copy of the k,v pair
-                if (!Array.isArray(kv))
-                    throw new Error(
-                        `VALUE ERROR key-value pair must be an array, entry ${i}: ${kv}`,
-                    );
-                const _kv = Array.from(kv);
-                if (!_kv.length >= 2)
-                    throw new Error(
-                        `TYPE ERROR Key-Value pair must a length ` +
-                            `of at least 2 [key, value] but entry ${i} ${_kv} length is ${_kv.length}`,
-                    );
-                const [valid, message] = this._validateKey(_kv[0]);
-                if (!valid)
-                    throw new Error(`TYPE ERROR in entry ${i}: ${message}`);
-                return Object.freeze(_kv);
-            }),
-            deleted = this._value.splice(
-                index,
-                deleteCount,
-                ..._entries.map((kvItem) => {
-                    const unwrapped = unwrapPotentialWriteProxy(kvItem[1]);
-                    if (kvItem[1] !== unwrapped) return [kvItem[0], unwrapped];
-                    return kvItem;
-                }),
-            );
-        // Replaces [index, key, proxy] by empty arrays, disconnecting proxies
-        this[_OLD_TO_NEW_SLOT].splice(
-            index,
-            deleteCount,
-            ...new Array(entries.length).fill(null).map(() => []),
-        );
-        // We can have duplicate keys in entries and we can have
-        // duplicate keys in this._value already.
-        const seen = new Set(),
-            deletedOnInsert = [];
-        for (let i = this._value.length - 1; i >= 0; i--) {
-            const kv = Object.hasOwn(this._value, i)
-                ? this._value[i]
-                : // Can use this[OLD_STATE].value because this[OLD_STATE]
-                  // is immutable.
-                  this[OLD_STATE].value[this[_OLD_TO_NEW_SLOT][i][0]];
-            const [key /*value*/] = kv;
-            if (seen.has(key)) {
-                // remove duplicate
-                deletedOnInsert.push(...this._value.splice(i, 1));
-                // also disconnect these proxies
-                this[_OLD_TO_NEW_SLOT].splice(i, 1);
-                continue;
-            }
-            seen.add(key);
-            if (!Object.isFrozen(kv))
-                // defensive copy
-                this._value[i] = Object.freeze(kv.slice());
-        }
-        // We iterated backwards, this is a better order.
-        deletedOnInsert.reverse();
-        deleted.push(...deletedOnInsert);
-        this._updateKeys();
-        return deleted;
-    }
-
-    splice(startKey, deleteCount, ...entries) {
-        const [index, message] = this.keyToIndex(startKey);
-        if (index === null) throw new Error(message);
-        return this.arraySplice(index, deleteCount, ...entries);
-    }
-
-    // This method will push undefined keys to the end.
-    set(key, newEntry) {
-        const [index /*message*/] = this.keyToIndex(key);
-        // replace or append
-        this.arraySplice(index === null ? Infinity : index, 1, [key, newEntry]);
-    }
-
-    delete(key) {
-        const [index /*message*/] = this.keyToIndex(key);
-        if (index === null) return;
-        return this.arraySplice(index, 1)[0];
-    }
-
-    // append, add to end
-    push(...entries) {
-        this.arraySplice(Infinity, 0, ...entries);
-        return this.size;
-    }
-    // add to front
-    unshift(...entries) {
-        this.arraySplice(0, 0, ...entries);
-        return this.size;
-    }
-    // remove from end
-    pop() {
-        return this.arraySplice(-1, 1)[0];
-    }
-    // remove from front
-    shift() {
-        return this.arraySplice(0, 1)[0];
-    }
-
-    /**
-     * It's like an "upsert"
-     * mapLike.constructor.Model must be a "simple model"
-     *       with a set(rawValue) method and without dependenies
-     *       required for Model.createPrimalDraft
-     * usage:
-     *   axisLocations.setSimpleValue(, axisTag, value)
-     */
-    setSimpleValue(key, value) {
-        if (!this.has(key))
-            this.set(key, this.constructor.Model.createPrimalDraft());
-        this.get(key).set(value);
-    }
-
-    [SERIALIZE](options = SERIALIZE_OPTIONS) {
-        return _serializeContainer(
-            this,
-            /*presenceIsInformation*/ true,
-            /*keepKeys*/ true,
-            options,
-        );
-    }
-}
+import { _AbstractOrderedMapModel, toShuffle } from './ordered-map-model.ts';
+export { _AbstractOrderedMapModel, toShuffle };
 
 //
 // has all of the dependencies OR one model that can change completely
@@ -1668,7 +814,7 @@ export function* getFieldsByType(FromType, SearchType) {
 
 export class Path {
     // jshint ignore: start
-    static SEPARATOR = "/";
+    static SEPARATOR = PATH_SEPARATOR;
     static RELATIVE = ".";
     static ROOT = "/";
     static PARENT = "..";
